requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
PyJWT==2.6.0
Werkzeug==2.2.3
click==8.1.3
python-dotenv==1.0.0
